refactor: rename/move item_forecast_config module and update import paths for ForecastItemConfig
