+++ finstmt/config/statement_config.py
@@ -4,7 +4,7 @@ import os
 from pathlib import Path
 
 from finstmt.findata.item_config import ItemConfig
-from finstmt.forecast.config import ForecastConfig, ForecastItemConfig
+from finstmt.findata.item_forecast_config import ForecastConfig, ForecastItemConfig
 from finstmt.config.config_loader import load_yaml_config
 
 @dataclass
